Checkpointing the current state of my work.
My current point of blockage is when we load fields, we need to pass
in FieldInfo objects.  This is easy for fields of built-in classes;
just use .GetField on the type object.

But you can't use .GetField on classes that you are in the midst of
constructing; you need to explicitly carry around the FieldInfo
(really FieldBuilder, but FieldBuilder extends FieldInfo) objects that
are returned by .DefineField.  This again seemed like it would be
easy; just keep an extra table directly mapping Type*FieldName ->
FieldInfo for the ones that we've constructed thus far.

This however is insufficient, because of *inheritance*!  If we define
a field "foo" for class A, and later query class B for "foo", and B
extends A, then we need to do the table lookup keyed on A.

To make matters more difficult, it doesn't look like the CLR provides
a member function mapping a type to its supertype.  Which I guess
makes sense for supporting languages like C++; it just means that this
code is going to get really hairy really quickly.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2569 c9681fa6-b8ae-40c0-a6da-707a35ad5973