Added an (extremely hackish) procedure to parse a path into a directory specifier prefix and a target file/directory.
This is necessary because the .NET libraries require
.DefineDynamicAssembly to take a storage directory as an argument; you
can't pass the directory to save to at a later point.  Thus, we need
to infer an appropriate directory to store to from the path for the
.fasl file we are generating.

(I would love to hear about alternate solutions to this problem)

(I would also love to make this implementation more robust.
 In particular, the sysdep-win32.sch version *requires* the path
 separators to be back-slashes.  It would almost certainly be
 better to interpret both back- and forward-slashes as directory
 separators.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2591 c9681fa6-b8ae-40c0-a6da-707a35ad5973