Added (complete hack) to allow me to lookup code stored in .dll files generated by Asm/IL/il-corememory.sch
I think that having Load.cs do this sort of guessing game is a mistake
though; it would probably be better to pass the full expected filename
as the argument to the segment_code_address syscall, and then have
that passed along to Load.cs.  Need to talk to Ryanc about this first.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2593 c9681fa6-b8ae-40c0-a6da-707a35ad5973