haha!  Yet another bug uncovered by the attempts to compile the compiler.
We almost never emit the lexes macscheme instruction; it is only
emitted for the code implementing pass3commoning.sch when compiling
the compiler.

However, the usage of this instruction revealed that we were checking
a stronger invariant than the MacScheme machine actually maintains.
We were assuming that Slot0 (of the topmost stack frame) and Reg0 have
to point to the same closure object.

Will explained to Joe and me that the invariant is actually that Slot0
and Reg0 have to point to closure objects which both point to the same
*codevector* object.  Will says this is to ensure that the codevector
is actually reachable when the garbage collector runs (the old program
counter will only point to some offset within the code vector, which
is not sufficent for the GC).

So, the point is that we can't assert that this.s0 == reg0 when we do
the pop; we can only assert (this.s0.entrypoint == reg0.entrypoint).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2595 c9681fa6-b8ae-40c0-a6da-707a35ad5973