Two changes:
 - Larceny is now path-aware.
 - Command line options have changed.

Larceny now attempts to figure out its location, which it can then use
to find support files.  Setting the environment variable LARCENY_ROOT to
a path will override this behavior and cause it to use that path rather
than trying to guess.

On Windows, it accomplishes this through an API call in
Rts/Sys/osdep-win32.c.  On UNIX, it accomplishes this through a shell
script, a prototype of which is in Scripts/petit in this commit.

In larceny.c, we now use the path stored in LARCENY_ROOT (either by the
user or by our guessing) to locate the default heap, if one is not
specified.

The heap is now specified on the command line using option -heap,
instead of specifying it as a bare option.  Also, -- may be used instead
of -args to set off program arguments, as this puts us more in line with
standard UNIX practice.

Let it be known that I need to start using branches, because this
shouldn't actually be a trunk commit.  I'm currently using this to
transfer code between home and the office.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2800 c9681fa6-b8ae-40c0-a6da-707a35ad5973