I've added a new system call, syscall:setenv at 46.  It takes two arguments, both strings and sets the environment variable named by the first argument to be the second argument.  A user function setenv is defined in system-interface.sch and exported in toplevel.sch.
I've also added another function in terms of getenv and setenv, called
make-env-parameter.  Given an environment variable name, it makes a
parameter that stays in sync with it.  For example:

    > (define current-shell (make-env-parameter "SHELL"))
    current-shell

    > (current-shell)
    "/bin/zsh"

    > (current-shell "/bin/sh")
    "/bin/zsh"

    > (current-shell)
    "/bin/sh"

    > (parameterize ((current-shell "foo")) (current-shell))
    "foo"

    > (current-shell)
    "/bin/sh"

TODO?: Add optional parser/unparser arguments, so that env parameters don't
have to be strings.  Or maybe just allow it to use read.  How do we read
from a string?



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2804 c9681fa6-b8ae-40c0-a6da-707a35ad5973