I'm pretty sure that these are strict improvements on what we had before.
The implementation of macro-expand that's given in this file can't
possibly work without a definition for macro-expand-expression.  What
worries me here is: Why did Lars comment it out in the first place?

The addition of the -heap option to the petit command at the end is to
accommodate Jesse's changes.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2859 c9681fa6-b8ae-40c0-a6da-707a35ad5973