Fixing bugs in the instruction generation.
A *lot* of the labels were left in quoted form.  Those are relatively
easy to see when you pretty print the output.

Also several of the "sassy macros" are no longer sassy macros but
instead are scheme procedures now, which means that they need to be
invoked, not quoted.

Fixed a place where I used the fictional "orr" instead of the
appropriate logior.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2877 c9681fa6-b8ae-40c0-a6da-707a35ad5973