Replaced all occurences of *_HDR with $hdr.***
I should probably do the occurences of *_TAG next (I caught some of
them this time, but did not try to replace all of them)

No new code builds yet, because it turns out that NASM has the ability
to dispatch based on the number of arguments in its macro system, but
I didn't realize that when I translated the file over, so there are
actually redefinitions of functions like: ia86.indexed_structure_ref,
which is no good.  Its not clear whether this was even intended on
Lars' end, but it certainly isn't going to work with my current setup.
So I'll probably disambiguate the overloading statically by hand.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2887 c9681fa6-b8ae-40c0-a6da-707a35ad5973