Replaced uses of G_THIRD with $g.third
Replaced IMM_* with $imm.***

Alpha-renamed some of the generic x/y/z names with the intended names
that Lars had in the comments.

Distinquished between the two variants of ia86.indexed_structure_ref
and ia86.indexed_structure_length (with a "/hdr" suffix), to get
around the fact that NASM add new variant definitions that dispatch
based on arg count, and Scheme (thankfully!) can not.

Added quotes to a bunch of the fixnum compare calls.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2889 c9681fa6-b8ae-40c0-a6da-707a35ad5973