Added a T_OP3 implementation.
Fixed a bug where I was over-eager in my addition of ia86. prefixes.

Jesse explained Intel's addressing to me: you can do, for example
{{{(mov (& R (* c R) c) ARG)}}} (where the R's are registers and the
c's are constants) and things will be happy.  You can also leave out
any of the three operands.  This led me to fix a number of places
where I had liberally replaced the expressions Lars had put in the
NASM source with something that Sassy is actually happy with (I was
trying to statically evaluate expressions that referred to the CONT
register, which is impossible)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2892 c9681fa6-b8ae-40c0-a6da-707a35ad5973