Fixed a few bugs at the expense of introducing a slight inconsistency with how the NASM backend is structured.
In NASM, stkslot(..) is a macro that expands into some sort of
arithmetic expression calculating a stack address.

In the Sassy backend, (stkslot ...) will expand into the actual
addressing operation, complete with the (& ...).  The alternative was
to give back a list of the operands and use ,@ everywhere, but so far
I've gotten away with only using ,@ once (thanks to the seqlist macro
hack), and I want to keep things that way.  

I'm not scared of unquote-splicing, but I do think this code is very
readable the way it is...


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@2906 c9681fa6-b8ae-40c0-a6da-707a35ad5973