Merged in changes from source:/branches/pnkfelix/iassassin/larceny_src@3081.
With these changes, one should be able to build a native larceny for
Intel using the Sassy assembler for the backend.
 * on win32, use the setup invocation 
   {{{> (setup 'scheme: 'larceny 'host: 'win32 'sassy)}}}
 * on linux, use the setup invocation 
   {{{> (setup 'scheme: 'larceny 'host: 'linux86 'sassy)}}}

'''However''': not all of the code in the iassassin branch was
transplanted directly.  In particular, the support code for
sassy-nasm-hybrid in Asm/Intel and Rts/Intel has been omitted, since
there is no longer a need (in principle) for a hybrid system; that was
merely a bootstrapping step.

It probably wouldn't be a terrible idea to factor the invocation
routines out of Rts/Intel/i386-instr.asm (the same way that
sassy-invoke.sch was factored out of sassy-instr.sch) to allow one to
easily put back the hybrid invocation routines.

If you're curious, the way I generated the starting point for this 
merge was with the command:

{{{
% svn merge \
   svn+ssh://sualocin.ccs.neu.edu/proj/will/pnkfelix/svn-archives/trunk/larceny_src@2935 \
   svn+ssh://sualocin.ccs.neu.edu/proj/will/pnkfelix/svn-archives/branches/pnkfelix/iassassin/larceny_src@3081
}}}

Some final notes:
 * This is not as speedy as I would like.  Assembling itself seems quite slow.
   See source:/branches/pnkfelix/iassassin-fewer-sassy-macros for one idea
   on how to improve this.
 * Compiling srfis 60 and 69 seems to be a good thing to speed up compile times.
   Compiling Sassy itself is probably also helpful.  Note that compiling
   srfi-69 is non-trivial since Larceny's require doesn't understand 
   YouWantItWhen (yet!).
 * There is lots of room for improvement in the generated code, but Will
   wants to have a release (perhaps internal only) with the intel native
   backend before we make further changes to optimize it.
 * The compat files are loading Sassy in all cases at the moment.  
   This should be changed so that we only load Sassy if we're in a sassy
   backend (an easy way to support this is to have the compat files export a 
   {{{load-sassy}}} procedure rather than doing the load themselves.).



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3082 c9681fa6-b8ae-40c0-a6da-707a35ad5973