Workaround for Ticket #92.
This does remove the immediate infinite loop I was seeing.  However,
it is not clear if this is the right fix.  

In particular, the semantics of collect with request_bytes==0 should
probably be to just collect, period.  

The real question is what should mc_morecore (in the millicode.c file
from Rts/Standard-C/ and Rts/IAssassin/) be passing as the
request_bytes argument then, as 0 is clearly wrong when the garbage
collector is allowed to return with no free core memory with such a
value for request_bytes.  

To determine this, we need to figure out what the semantics of
mc_morecore are supposed to be, which we will probably have to infer
from how it is used.  (Then we'll correct the mc_morecore
implementation and its clients in tandem).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3104 c9681fa6-b8ae-40c0-a6da-707a35ad5973