Bug fix for #178.
Mac OS X doesn't have the timezone global variable available, so we
need to be a little smarter and calculate the timezone offse based on
what standard in the libraries.

I discovered this technique in the time.c code of the Apache Portable
Runtime (APR).  The insight is simple; mktime o localtime should be
the identity function, so
 (lambda (x) (- (mktime (localtime x))
                (mktime (gmtime x)))) 
should yield the offset at a particular point in time.

But one worrisome point is that the APR developers seemed concerned
about DST; apparently gmtime was turning it on (bizarre) and so they
had to explcitly turn it off and perahps even correct by an hour.  

But even with the potential problem there, this is still much better
than what we had before, at least on Mac OS X.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3743 c9681fa6-b8ae-40c0-a6da-707a35ad5973