Revised the code that checks for references to undefined labels so that it does not need the list of all opcodes that might be part of the generated assembly sequence.
The code now traverses the s-exp with some knowledge about what it is
getting (that the input is a list of instructions) and therefore can
skip over opcodes instead of treating them as references to
identifiers that need to be looked up.

This eases the introduction of new primitives and of inline assembly
sequences that use machine instructions while the rest of larceny
remains oblivious.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3797 c9681fa6-b8ae-40c0-a6da-707a35ad5973