Added a slight hack to our run script to make it easier for me to invoke Larceny under a gdb session.
However, this is not really the right way to code this up, because the
choice of which Larceny executable and heap to run is orthogonal to
whether we should host the runtime atop gdb.

So someone in the future should revise the case statement in the
switch to match the choices more liberally (e.g. match *-larceny,
*-larceny-*, larceny-*, and larceny for the larceny case), and then we
will be able to mix in things like turning on gdb (via matching
*-gdb-* and gdb-*)

(I believe it is only the matching code that is limited in this
fashion; the rest of the addition should be generic enough to allow
one to use gdb to debug Petit and/or Twobit at a whim.)



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3839 c9681fa6-b8ae-40c0-a6da-707a35ad5973