Whatever I had here before did not have the right semantics. It was doing a jump to *fptr instead of just to fptr.
I could not quickly figure out how to patcht a direct call to an
address into the machine code we generate in the tr-at-end call.  So
instead of working out the details now, I do a mov of the function
pointer into eax and then a call to eax, just to get things working.
(This is not efficient.)

I need to double-check that it is legal for me to blow away eax in
this fashion.  I know that the callee can return values in eax, so it
certainly is a caller save register (which we have no need to save),
but I'm not certain if the parameter conventions allow for one of the
arguments to be passed via eax.  If so, then this *has* to be revised
for correctness.



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3840 c9681fa6-b8ae-40c0-a6da-707a35ad5973