Jed suggested this change to pass3's {{{anf-order-of-evaluation}}}.
He explains the philosophy behind it in the following message to the
larceny-users mailing list:

https://lists.ccs.neu.edu/pipermail/larceny-users/2007-February/000025.html

I benchmarked the effects of the change, and posted the results here:

http://www.editgrid.com/user/pnkfelix/jed-anf-less-complicated.html

In a nutshell, it helps Intel a little bit (1%-2% overall) and does
not do much on Sparc. The times it helps on Intel, it *really* helps.



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@3999 c9681fa6-b8ae-40c0-a6da-707a35ad5973