Changed code to use named offsets (rather than magic numbers).
Note that this does not yet fix #401 because the assembly code still
assumes that writes to the frame will not clobber values of other
offsets (which is no longer true when the retaddr and contsize are
swapped).  The SCE_BUFFER (that we really only need on IasnLarceny)
definition might also be a problem; I am not sure about that yet.




git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4005 c9681fa6-b8ae-40c0-a6da-707a35ad5973