Fix for latent bug in sparc backend.
While the inlined assembly sequence would save and restore REG0 when
restoring a frame (to support Will's multiple values code) the backup
code that called out to C (which you can explicitly use by changing a
#if 1 to #if 0 at the start of mem_stkuflow) did not do this.

This is actually very tricky, because calling out to C will attempt to
convert the return address to an offset into the bytevector, based on
the closure pointed to by REG0.  But wait, REG0 might not be a
closure!  The callout code already considers this a possibility; the
exceptional condition is signalled when REG0 holds the value 0.

(But of course with the new multiple values protocol, REG0 might be
any fixnum, not just 0.)

So when I save REG0 in the callout code for mem_stkuflow, I check if
REG0 is a fixnum.  If it is, then I store 0 in REG0.  (We'll be
restoring REG0 anyway, so its safe to clobber it with 0.)  Then the
callout code won't attempt to adjust the return address.




git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4006 c9681fa6-b8ae-40c0-a6da-707a35ad5973