I was too aggressive with changeset:4070; almost all the builds failed last night.
This is a fix for the non IA32 runtimes.  (I will address win32 in a different commit.)

Here is what changeset:4070 was trying to fix: when we increase a
semispace's {{{.top}}}, we need to ensure that any uninitialized words
are ignored.  The way that we do this in {{{seal_chunk}}}
([source:trunk/larceny_src/src/Rts/Sys/cheney.c cheney.c]) is to put a
dummy object in, a string that covers the entire space that is
uninitialized.  So changeset:4070 was attempting to do the same thing,
putting in a string header that covered the space from {{{.top}}} to
{{{.lim}}}.

Here is what was wrong with changeset:4070: the uninitialized space is
not the area from {{{.top}}} to {{{.lim}}}.  At least, not for the
current allocation area, because while the nursery grows from
{{{.bot}}} to {{{.top}}}, the ''stack'' grows from {{{.lim}}}
downward.  So the uninitialized space is the area from {{{.top}}} to
the top of the stack ({{{globals[G_STKP]}}}).

So I've fixed the code so that it calculates based on the limit being
the end of the stack, not the whole chunk's {{{.lim}}}.

I also made a small fix to not put the 0xABCDABCD marking bit pattern
in the "padding" unless we actually have space for it (though the
chances of this actually mattering in practice are pretty small...)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4074 c9681fa6-b8ae-40c0-a6da-707a35ad5973