Bugfix for #424.
A buggy attempt to eliminate redundant moves was introduced in
changeset:3514.

The move elimination was buggy because I did not properly consider
that in the generated assembly, control could branch to the label
{{{l1}}} from the typetag test at the start of the generated assembly
(namely when the input is not a fixnum).  When control hits the
aforementioned typetag test, register {{{rs1}}} has not yet been moved
into register {{{rd}}}, and so we cannot claim to have established an
invariant that when {{{rd}}} is the result register, it is not
necessary to move {{{rs}}} into the result before calling out to the
millicode arithmetic routines.

The unfortunate thing about this fix is that fixnum arithmetic
overflows will move {{{rs1}}} into result twice when {{{rd}}} is the
result register, which is a common case.  (That was why I put the move
elimination in in the first place.)  We may want to reconsider the
control structure of the code generated here to see if there is a way
to avoid that.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4241 c9681fa6-b8ae-40c0-a6da-707a35ad5973