This changeset addresses #429.  Common Larceny now uses ustrings rather than strings as its representation of Scheme strings.
We replaced calls to Factory.makeString with Factory.makeUString in
several places without thinking too carefully about the implications,
because it seemed reasonable.  (The only place we didn't change is for
the currently unused make_string primop.)  We aren't sure how exactly
this will affect the FFI or Factory.wrap, for example.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4451 c9681fa6-b8ae-40c0-a6da-707a35ad5973