Bug fix to {{{g-signal-connect-data}}}; when we get a callback that returns {{{void}}}, we should not treat that as a {{{void*}}}.
(This was breaking in a clean way, since the callback in this case was
returning {{{#!unspecified}}} and the marshalling code did not know how to
convert {{{#!unspecified}}} to a {{{void*}}}.)

The strangest thing about all this was discovering that glib is
claiming that some higher order functions return `{{{v }}}' instead of
`{{{void}}}' or `{{{v}}}'.  What's up with that?


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4454 c9681fa6-b8ae-40c0-a6da-707a35ad5973