Addressing Ticket #400.
Interruptible syscalls, when interrupted, will return with the
{{{UNDEFINED_CONST}}} in the result register (see
{{{BEGIN_INTERRUPTIBLE_SYSCALL}}} in [source:trunk/larceny_src/src/Rts/Sys/signals.h signals.h]).

There are different ways one might choose to handle such a result, but
blithely attempting to apply numeric operators to it is not
appropriate.  For now, I am returning {{{'error}}} from the read and
write requests, and that will be handled by the IO system like any
other exception.  But this may not be the way we want to do things in
the future; see the comments with Ticket #400 for further discussion.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4528 c9681fa6-b8ae-40c0-a6da-707a35ad5973