Added a crazy amount of code to support an easy way to declare and marshall "out" parameters in a simple way for client code.  Such parameters are currently declared using a pseudo-type-constructor ("pseudo-" because it really isn't a type constructor; its just sugar that means something special when it occurs in argument position in a {{{define-foreign}}} form, and in other contexts it is meaningless)
Here is some sample code to demonstrate how to use this feature

{{{
(require 'foreign-sugar)

(call-with-output-file "foo.c"
  (lambda (out)
    (display 
     "int foo(int x, int *y, int z) { *y = x + z; return x - z; }" 
     out)))

(system "gcc -c foo.c -o foo.o")
(system "gcc foo.o -flat_namespace -bundle -o foo.so")
(foreign-file "foo.so")

(define-foreign (foo int (ret int) int) int)

;; Invoking (foo 1 2 3) fails; it only takes 2 args.
;; Invoking (foo 2 1) returns *two* values: 2-1 (=1) and 2+1 (=3)

;; Note that the main return value comes first; this is nice when you
;; have the semantics that extra returned values are silently dropped.
}}}

The nice thing about the {{{(ret X)}}} feature is that it saves the
user from the trouble of manually writing the code to create a box and
pass it into the function expecting an {{{X*}}}.  

On the other hand, the code as it stands only properly marshalls
certain particular {{{X}}}'s (e.g. the stored value must be
word-sized, and it can't require any special handling beyond a
straight-forward %peek32).  It is not impossible to generalize things
to support {{{double}}}s and {{{float}}}s, but I do not want to
struggle with that change at the moment.  The macro does work on
{{{int*}}} and {{{char**}}}, which are the two common cases I want to
handle at the moment.

The macro is written in a pretty unsafe manner; in particular I am not
using the rename procedure {{{ren}}} at all, so it is not very robust.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4552 c9681fa6-b8ae-40c0-a6da-707a35ad5973