Added {{{save/storem}}} and {{{loadm}}} peepholes to reduce codesize on CommonLarceny.
Reducing code size improves JIT compilation time: the (real) time to
start and then immediately exit CommonLarceny goes from 320 seconds to
213 seconds.

Note that the majority of the time, and the majority of the reduction,
is due to time spent JIT compiling the R6RS-compatible reader.

Before these peepholes, Mono's profiler said that it spends 376.1
seconds compiling 28582 methods, and of that, 301.6 secs is spent
compiling the reader (80.2% of the time).

With these two peepholes, the profiler says that Mono spends 269.5
seconds compiling 28582 methods, and of that, 195.3 secs is spent
compiling the reader (72.5% of the time).

There is almost certainly other low-hanging fruit to peephole within
the reader.  (But beware of diminishing returns...)



git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4696 c9681fa6-b8ae-40c0-a6da-707a35ad5973