Fix for Ticket #468.
The autobuild system deliberately tries to avoid entering a host
debugger.  It communicates this desire by defining the global variable
{{{*exit-on-error*}}} to {{{#t}}} before loading the compatibility
file for the host Scheme system.

Then, the compatibility file is expected to respond to this by doing
whatever it needs to make an invocation of the {{{error}}} procedure
cause the host Scheme system to exit with a non-zero error code
status.

When running on PLT Scheme versions prior to v370, we accomplished
this by overriding the {{{current-exception-handler}}} parameter.
With the changes made in v370, we have to change our approach and
instead override the {{{error-escape-handler}}} parameter (which has a
somewhat different semantics, but will serve fine for our purposes).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4700 c9681fa6-b8ae-40c0-a6da-707a35ad5973