Further changes to address Ticket #169.
Mono is stricter than Microsoft about enum values; it seems like Mono
always boxes enums and checks that the type tags match, while
Microsoft's runtime was accepting integers as long as they matched the
enum int values.

This led to Felix writing code in this file that worked on Microsoft's
runtime but caused an exception on Mono's runtime.  Now it seems like
Mono can compile code in core memory, which means that Ticket #169
might be resolved.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4714 c9681fa6-b8ae-40c0-a6da-707a35ad5973