Fixed this code so that pre-existing format flags are not overwritten when I add in the {{{'measuretrailingspaces}}} flag.
Also, I've now determined that this code does seem to properly measure
spaces on Microsoft's .NET implementation; Mono seems to have a bug.

It would be relatively easy to hack around; just add a "a" to the end
of the string and then subtract the width of "a" from the final result
width.  Maybe I will add that tomorrow morning.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4745 c9681fa6-b8ae-40c0-a6da-707a35ad5973