Shifted rendering of graphics to a sub-panel within the form rather than drawing on the form's graphics object directly.
This required a bit more code than you would expect because I had to
make the sub-panel double buffered.  So I abstracted the reflective
code that generates a fresh class and constructor (that sets the
protected {{{DoubleBuffered}}} property to {{{true}}}) over the type
the new type is extending.  This makes it easy to switch back to the
approach I was using before (where I drew the graphics directly on the
form object itself).

The motivation for using a sub-panel rather than drawing directly on
the form is to allow the separation of input event handling for
scrollbars from the rendered sub-panel, but this separation has not
actually been implemented yet (input event handling is still performed
by the form itself, rather than the sub-panel).  Such a separation of
the event handling is necessary because with the current state of
things, a down keyevent is sometimes treated as both a downward
movement of the cursor and as a movement of the scrollbar; the only
time that should happen is when the cursor is at the last visible line
in the window, but right now it even happens when the cursor is on
other lines in the window.

(Such a separation might not be the only way to handle this problem;
there are some event argument types (e.g. {{{KeyPressEventArgs}}})
that have a {{{Handled}}} property that indicates that the event
should not be fed into containing controls.  But this is not true of
event argument types in general, so this strategy would seem to be
inapplicable in this particular use case.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4767 c9681fa6-b8ae-40c0-a6da-707a35ad5973