Panels do not seem to handle key events, so I switched to extending {{{Control}}} directly for the form contents instead of extending {{{Panel}}}.
Also, while debugging this code I noted that I did turn the
{{{KeyPreview}}} property to {{{true}}} earlier, but was not setting
the {{{Handled}}} property to {{{true}}} when handling events within
the form (indeed, it is hard to do this well because I would probably
want to have the agent communicate whether or not the event should be
propagated to the subcontrol).  

The point is that the inconsistency I was struggling with (that the
scrollbar and the cursor were both responding to input key events) may
have been resolved more simply by playing with the {{{Handled}}}
property.  But at this point it seems like this approach of
introducing a sub-panel (and attaching the events to that instead of
to the top level form object) works well enough that I do not need to
investigate the {{{Handled}}} property (at least for now).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4768 c9681fa6-b8ae-40c0-a6da-707a35ad5973