Started work on revising code.
The goal of the revision is to enforce a split between the text view
and the backing code editor.

The text view is meant to hold ''visible'' text and interprets simple
keyboard and mouse events.

The backing code editor will hold the entirety of the text file and
will override some of the handler functionality to respond to events
in more sophisticated ways (e.g. by auto-indenting on a Enter key).


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4848 c9681fa6-b8ae-40c0-a6da-707a35ad5973