Change so that we don't get an error if we attempt to move the cursor beyond the first line.
This is only a temporary patchup -- the long term solution will be to
see if the backing agent has more content above the current line, and
if so, issue a scroll event, and then make a retry on moving the
cursor.  But even with that change, in the limit I will still need to
handle this case where there is no more content to be seen, in which
case making the key up event a no-op is probably still the right
behavior.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4852 c9681fa6-b8ae-40c0-a6da-707a35ad5973