Changed mouse event processing code so that the end of a line is treated as extending infinitely far to the right.
It would be nice to also change things so that the final line is
treated as having infinite extent below and the first line has
infinite extent above, but that could require feeding a bit more
context information to the {{{point-in-range?}}} function.  Well,
that, or some additional special cases in the mouse event processing
code.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4854 c9681fa6-b8ae-40c0-a6da-707a35ad5973