Discovered that this interface does not provide for one to properly indent a named let.
I'm still thinking about the "right" way to handle this, but the
easiest change seems to be to feed more information about the
{{{to-first}}} and {{{to-final}}} arguments to the
{{{indent-suggest}}}; the question is what form should this
information take?  The maximally expressive approach would be to feed
the whole s-exp through, but that would complicate the problem more
than I want to right now.  Another approach, which I am currently
adopting, is to just pass along the character that starts off the
{{{to-first}}} (and {{{to-final}}} argument; that will be enough to
distinquish between a named let and a normal let.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4896 c9681fa6-b8ae-40c0-a6da-707a35ad5973