Figured out an effective way of using the magic of {{{syntax-rules}}} ellipses to propagate the arity of the method argument list into the methods we return.
This does not propagate any arity information for methods with
variable arguments (but I am specifically avoiding adding such methods
to the interfaces of my toolkit).  I wouldn't mind learning the
appropriate pattern to do this sort of trick for that case as well.

An added feature of the macro is that it incorporates procedure-name
information, which simplifies the client code in
{{{make-root-object}}}.

The main goal was to make the procedure documentation for the objects
and methods as informative as possible.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4906 c9681fa6-b8ae-40c0-a6da-707a35ad5973