Added support for inline method documentation.
You pull out the documentation for a method using the
{{{'documentation}}} method, passing the operation symbol along to the
method.

One trick is that you will "inherit" documentation for methods by
leaving out the documentation string.  As an example, the
{{{colored-point}}} extension of {{{point}}} does not document its
redefinition of the {{{move}}} factory method, so an attempt to find
the documentation of {{{'move}}} of a {{{colored-point}}} passes the
buck along to the {{{point}}} root object.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4943 c9681fa6-b8ae-40c0-a6da-707a35ad5973