Now looking at the REPL again.
Hopefully the new refactored architecture will make it easier for me
to add auto-indenting and paren-matching etc support to the REPL.  But
my focus in the short term is just on getting things to the same state
they were in before I started the refactoring yesterday.

Curiously enough, very few changes were necessary to get things to a 
semblance of working.

  1. I renamed the {{{editor-agent}}} formal variable to 
  {{{textmodel}}}, just to be intellectually honest... :)
  
  2. I changed the {{{repl-agent}}} construction from a 
  root object construction to an extension of the {{{textmodel}}}.

  3. I changed various references to {{{editor-agent}}} (aka
  {{{textmodel}}}) to the self reference {{{repl-agent}}}
  (which, as noted in (2) above, extends {{{textmodel}}} anyway).

Given how small the above three changes were, I'm a little bit curious
as to whether this code would have "just worked" as is without any
changes at all.  (But not curious enough to try actually doing it; 
lets move forward.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4964 c9681fa6-b8ae-40c0-a6da-707a35ad5973