Refactored code so that the output port for the REPL is in scope for use by the reader.
This is to enable reparameterization the current output port so that
macro expansion warnings (e.g. of expressions like {{{()}}}) can be
printed to the REPL window.  However, the obvious way to do this did
not seem to work when I tried it, so there's a bit more to be done
here...


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4984 c9681fa6-b8ae-40c0-a6da-707a35ad5973