Mixed paren matching into the REPL code.
Arrow keys are not yet supported in the REPL (only the editor), which
makes paren matching less useful than I would like.  But it is still
semi-useful.  

I do not know how high a priority it is to make arrow keys work; it
might be best if they stayed disabled for the time being.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4989 c9681fa6-b8ae-40c0-a6da-707a35ad5973