Ah, the joy of adding features via code simplification!
I added arrow key handling support to the REPL by changing how work is
distributed in the REPL's methods.  Now, the only key stroke that the
REPL handles directly is Enter; otherwise it delegates to the
underlying agent.  The filtering of attempts to edit text before the
prompt is now handled by overriding the {{{insert-char-at-point!}}}
and {{{delete-char-at-point!}}} methods.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@4991 c9681fa6-b8ae-40c0-a6da-707a35ad5973