Some fixes to scrollbar handling.
When we automatically scroll in response to a cursor reposition, we
move the buffer by half of the visible lines in the buffer.  The
intention is that when the cursor moves past the end of the visible
area, we'll shift the visible area so that the cursor is now on the
center line.

The problem was that I was unconditionally calculating an amount to
scroll by, and then attempting to scroll by that amount, even when we
*hadn't* moved the cursor past the end of the visible area.  This led
to oddness where, e.g., if you had a single line and hit the enter
key, the single line would scroll out of view and the cursor would
appear at the top of the visible area.

This change attempts to counteract this for the scrolling forward case
by calculating the number of actual lines in the visible text (as
opposed to the number of lines that we could hypothetically display if
we managed to fill in some content), and then it only attempts to
scroll when we have more content than space.  (I don't think there's
anything to fix for the scrolling backward case.)


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5000 c9681fa6-b8ae-40c0-a6da-707a35ad5973