Changed rendering code so that we draw the entirety of the background followed by the entirety of the foreground.  This yields better rendering of things like #\0
Explanation: for some reason on Mono at least, the right side of the
zero seems to go past the calculated width of the zero, which leads
the previous background rendering code to erase part of the zero it
just drew.  The same held for other characters like {{{#\Q}}}, 
{{{#\0}}}, {{{#\U}}}, {{{#\M}}}, {{{#\N}}} and many others.

Another fix for this problem (and some other problems) would be to
render larger swathes of graphical objects; instead of a character at
a time, render as much of a string as I can (that is, whatever portion
of a line that has uniform selection state).  I might tackle that
next.


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5003 c9681fa6-b8ae-40c0-a6da-707a35ad5973