Making the tests run by the nightly build script match up with the tests that are run by [source:trunk/larceny_src/test/Lib/load-all-tests.sch load-all-tests.sch].
(I had thought that only the hashtable tests were left to be added,
but Will has been very busy adding files to our test suite.)

In an ideal world we would have a level of indirection somewhere so
that these files did not have to be kept in sync with each other.  Or
maybe not; maybe letting [source:trunk/larceny_src/test/Lib/load-all-tests.sch load-all-tests.sch] 
be relatively free form while the autobuild script is kept stable is a
better approach. . .


git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@5053 c9681fa6-b8ae-40c0-a6da-707a35ad5973